# Optimize array allocation.