--- a/lib/cudadrv/pool.jl
+++ b/lib/cudadrv/pool.jl
@@ -78,8 +78,8 @@ trim(pool::CuMemoryPool, bytes_to_keep::Integer=0) = cuMemPoolTrimTo(pool, bytes
 Returns attribute `attr` about `pool`. The type of the returned value depends on the
 attribute, and as such must be passed as the `X` parameter.
 """
-function attribute(X::Type, pool::CuMemoryPool, attr::CUmemPool_attribute)
-    value = Ref{X}()
+function attribute(::Type{T}, pool::CuMemoryPool, attr::CUmemPool_attribute) where T
+    value = Ref{T}()
     cuMemPoolGetAttribute(pool, attr, value)
     return value[]
 end